update: automate curl instruction check
update: automate curl instruction check

runInstall: use options to bypass gui

update: automate curl instruction check

update: automate curl instruction check

update: automate process to check instructions work

+++ install.bat
@@ -4,6 +4,26 @@ REM cygwinInstall
 
 cd /D "%~dp0"
 
+:: Parameters for running scheduled task
+set "_parOneCurlCygwinInstall=%~1"
+set "_checkParOneCurlCygwinInstall=-%_parOneCurlCygwinInstall%-"
+
+:: Use `_parOneCurlCygwinInstall` with --delay for delayed task to check install "Pass" or "Fail"
+if "%_parOneCurlCygwinInstall%"=="--delay" (
+ taskkill /F /FI "imagename eq WindowsSandboxServer.exe"
+ if EXIST "sandbox" (
+  if EXIST "sandbox\curlInstructionWork.txt" (
+   move "sandbox\curlInstructionWork.txt" "curlInstructionWork.txt"
+  ) else (
+   echo Fail> curlInstructionWork.txt
+  )
+  rmdir /S/Q sandbox >nul 2>nul
+ ) else (
+  echo Fail> curlInstructionWork.txt
+ )
+ goto _removeBatchVariables
+ goto:eof
+)
 :: Process to ensure path is correct for config.
 cd> tmpFileCygwinInstall.txt
 sed -i "s/\\/-:-/g" tmpFileCygwinInstall.txt
@@ -22,52 +42,52 @@ call :_startCygwinInstall 1 & goto:eof
  if "%1"=="1" (
   rem update with current path
   sed -i -E -e "s/(<HostFolder>)(.*)(<\/HostFolder>)/\1%_curDir%\\sandbox\3/" -e "s/-:-/\\/g" runStartSandbox.wsb
-  
+
   rem update with config option
   sed -i -E "0,/(_configOption=).*/{s/(_configOption=).*/\1%_configOption%/}" sandbox\current.bat
   call instructLine /B
   call instructLine /H "INSTRUCTIONS:"
   TIMEOUT /T 1 >nul 2>nul
   call instructLine /B
-  call instructLine "When Sandbox Opens, open the mapped folder and double click "runInstall.bat"."
+  call instructLine "When Sandbox Opens, runInstall.bat will execute automatically."
   call instructLine /B
-  TIMEOUT /T 5 >nul 2>nul
+  TIMEOUT /T 3 >nul 2>nul
   call instructLine /H "CONFIGURE SANDBOX:"
   TIMEOUT /T 1 >nul 2>nul
   call instructLine /B
-  call instructLine "Do you also want to install Notepad++ for debugging in Sandox?"
-  call instructLine "input - yes or no"
-  call instructLine /B
-  set /P _installNotepad=
-  call :_checkInstallNotepad 1
-  call instructLine /B
-  
+  rem check if installing notepadd++ for debug in sandbox
+  if NOT "%_parOneCurlCygwinInstall%"=="--task-run" (
+   call :_checkInstallNotepad 1
+  ) else (
+   rem set Sandbox to shutdown
+   sed -i -z -E "s/set \"_runAsScheduledTask=[0-9]/set \"_runAsScheduledTask=1/" sandbox\runInstall.bat
+  )
   rem start sandbox, mapping clean sandbox folder
   runStartSandbox.wsb
-  
   rem remove tmp file for windows command line variable
   del tmpFileCygwinInstall.txt
-
   rem next part of process
   call :_startCygwinInstall 2 & goto:eof
  )
  if "%1"=="2" (
   rem output final notes on process
-  call instructLine /B
-  call instructLine "Wait for Sandbox Process to Finish, then Press Enter to Close this Window:"
-  call instructLine /B
-  TIMEOUT /T 4 >nul 2>nul
-  call instructLine /H "IMPORTANT"
-  call instructLine "NOTE - closing after Sandbox session has ended will remove temp files of build." 
-  call instructLine "NOTE - the entire install process should take around 15 minutes."
-  call instructLine "NOTE - if needed delete runStartSandbox.wsb and sandbox folder after install."   
-  call instructLine /B
-  call instructLine /D
-  call instructLine /B
-  pause
-  call instructLine "Delay of 15 seconds to ensure Sandbox task is closed."
-  Timeout 15 >nul 2>nul
-  call instructLine /B
+  if NOT "%_parOneCurlCygwinInstall%"=="--task-run" (
+   call instructLine /B
+   call instructLine "Wait for Sandbox Process to Finish, then Press Enter to Close this Window:"
+   call instructLine /B
+   TIMEOUT /T 4 >nul 2>nul
+   call instructLine /H "IMPORTANT"
+   call instructLine "NOTE - closing after Sandbox session has ended will remove temp files of build."
+   call instructLine "NOTE - the entire install process should take around 15 minutes."
+   call instructLine "NOTE - if needed delete runStartSandbox.wsb and sandbox folder after install."
+   call instructLine /B
+   call instructLine /D
+   call instructLine /B
+   pause
+   call instructLine "Delay of 15 seconds to ensure Sandbox task is closed."
+   Timeout 15 >nul 2>nul
+   call instructLine /B
+  )
   rem next part of process
   call :_startCygwinInstall 3 & goto:eof
  )
@@ -84,7 +104,7 @@ call :_startCygwinInstall 1 & goto:eof
    rem remove variables for process
    goto _removeBatchVariables
   )
-  
+
   exit /b
  )
 goto:eof
@@ -95,31 +115,39 @@ goto:eof
   rem remove prior runStartSandbox if exists
   if EXIST runStartSandbox.wsb del /Q runStartSandbox.wsb >nul 2>nul
   if "%2"=="--start" copy /Y StartSandbox.wsb runStartSandbox.wsb >nul 2>nul
-  
+
   rem remove prior sandbox if exists
-  if EXIST sandbox rmdir /S/Q sandbox >nul 2>nul
+  if NOT "%_parOneCurlCygwinInstall%"=="--task-run" (
+   if EXIST sandbox rmdir /S/Q sandbox >nul 2>nul
+  )
   if "%2"=="--start" (
    xcopy /E/I map sandbox >nul 2>nul
    copy /Y cmdVar.bat sandbox\cmdVar.bat >nul 2>nul
    copy /Y instructLine.bat sandbox\instructLine.bat >nul 2>nul
   )
-  
+
   rem if starting procedure, give option to specify config option
   if "%2"=="--start" (
-   rem select configuration option
-   call instructLine "Enter Corresponding DIGIT to Select Config Option:"
-   call instructLine /D
-   call instructLine "NOTE - press enter to use default --without-ssl option."
-   call instructLine /B
-   call instructLine /F config-options.txt
-   call instructLine /B
-   set /P _configOption=
-   
-   rem store number of lines in a variable
-   FOR /F %%A in ('find /v /c "" ^< config-options.txt') DO set _numberOfOptions=%%A
-   
-   rem step 2 - allow batch to process variable change
-   call :_prepForNewRun 2 & goto:eof
+   if "%_parOneCurlCygwinInstall%"=="--task-run" (
+    rem use default
+    set "_configOption=--without-ssl"
+    call :_prepForNewRun 2 --set-default & goto:eof
+   ) else (
+    rem select configuration option
+    call instructLine "Enter Corresponding DIGIT to Select Config Option:"
+    call instructLine /D
+    call instructLine "NOTE - press enter to use default --without-ssl option."
+    call instructLine /B
+    set /P _configOption=
+    call instructLine /F config-options.txt
+    call instructLine /B
+
+    rem store number of lines in a variable
+    FOR /F %%A in ('find /v /c "" ^< config-options.txt') DO set _numberOfOptions=%%A
+
+    rem step 2 - allow batch to process variable change
+    call :_prepForNewRun 2 & goto:eof
+   )
   ) else (
    rem ensure curl was moved
    if EXIST "curl" (
@@ -135,22 +163,27 @@ goto:eof
   )
  )
  if "%1"=="2" (
-  rem define config option per input
-  if NOT DEFINED _configOption (
-   set "_configOption=--without-ssl"
+  if "%2"=="--set-default" (
+   rem using scheduled task
+   call instructLine "Configuration set to %_configOption%"
   ) else (
-   rem ensure that correct digit was input
-   echo %_configOption% | findstr /R [1-%_numberOfOptions%] >nul 2>nul
-   if "%ERRORLEVEL%"=="0" (
-    rem store appriopriate option in variable from input digit
-    type config-options.txt | findstr "%_configOption%" | sed -E "s/^(%_configOption:~0,1%%)(.*)$/\2/" > _tmp-config-opt.txt
-    call cmdVar "type _tmp-config-opt.txt" _configOption
-    
-    rem remove temp file
-    del /Q _tmp-config-opt.txt >nul 2>nul
-   ) else (
-    call instructLine "Incorrect Input - Using Default --without-ssl"
+   rem define config option per input
+   if NOT DEFINED _configOption (
     set "_configOption=--without-ssl"
+   ) else (
+    rem ensure that correct digit was input
+    echo %_configOption% | findstr /R [1-%_numberOfOptions%] >nul 2>nul
+    if "%ERRORLEVEL%"=="0" (
+     rem store appriopriate option in variable from input digit
+     type config-options.txt | findstr "%_configOption%" | sed -E "s/^(%_configOption:~0,1%%)(.*)$/\2/" > _tmp-config-opt.txt
+     call cmdVar "type _tmp-config-opt.txt" _configOption
+
+     rem remove temp file
+     del /Q _tmp-config-opt.txt >nul 2>nul
+    ) else (
+     call instructLine "Incorrect Input - Using Default --without-ssl"
+     set "_configOption=--without-ssl"
+    )
    )
   )
  )
@@ -164,6 +197,14 @@ goto:eof
 
 :_checkInstallNotepad
  if "%1"=="1" (
+   call instructLine "Do you also want to install Notepad++ for debugging in Sandox?"
+   call instructLine "input - yes or no"
+   call instructLine /B
+   set /P _installNotepad=
+   call :_checkInstallNotepad 2 & goto:eof
+ )
+ if "%1"=="2" (
+  call instructLine /B
   if /i "%_installNotepad%"=="yes" (
    echo yes> "sandbox\install_notepad.txt"
   )
@@ -177,6 +218,8 @@ goto:eof
  set _curDir=
  set _configOption=
  set _numberOfOptions=
+ set _parOneCurlCygwinInstall=
+ set _checkParOneCurlCygwinInstall=
  call instructLine "COMPLETE:"
  call instructLine /B
  call instructLine /D
